requirements.txt
aiogram
aiohttp
aiosqlite
google-generativeai
colorama